enable build on macOS
